+++ lib/let_me/all_of.ex
@@ -1,13 +1,12 @@
 defmodule LetMe.AllOf do
   @moduledoc """
-  Struct that represents a combination of checks that all must be true.
+  Struct that represents a combination of expressions that all must be true.
+
+  An `AllOf` without children evaluates to `true`.
   """
 
-  alias LetMe.Check
-  alias LetMe.Literal
-
   @type t :: %__MODULE__{
-          children: [Check.t() | Literal.t()],
+          children: [LetMe.expression()],
           passed?: boolean | nil
         }